refactor(api): endpoints 규격에 맞게 변경
